backend/requirements.txt
bidict==0.21.2
certifi==2020.12.5
cffi==1.14.6
chardet==4.0.0
click==7.1.2
cryptography==3.4.7
dnspython==1.16.0
dominate==2.6.0
eventlet==0.30.2
Flask==1.1.2
Flask-Bootstrap4==4.0.2
Flask-Cors==3.0.10
Flask-Script==2.0.6
Flask-SocketIO==4.3.2
gevent==21.1.2
greenlet==1.0.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
pycparser==2.20
pyOpenSSL==20.0.1
python-engineio==3.14.2
python-socketio==4.6.1
requests==2.25.1
six==1.15.0
socketIO-client==0.7.2
socketIO-client-2==0.7.5
urllib3==1.26.4
visitor==0.1.3
websocket-client==0.58.0
Werkzeug==1.0.1
zope.event==4.5.0
zope.interface==5.3.0
